requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
